Abstract away main functionality, add LS3 currencies too
