+++ modules/index.js
@@ -2,5 +2,6 @@ export default [
     require('./dungeon-collection').default,
     require('./mastery-point-achievements').default,
     require('./sab-collections').default,
+    require('./ls3-currencies').default,
     require('./ls4-currencies').default
 ];
